[udpate]delete処理にPolicyの検閲を追加

+++ app/Http/Controllers/PostController.php
@@ -109,6 +109,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize("delete", $post);
         $post->delete();
         return redirect("/games");
     }
